backend/requirements.txt
Flask==2.2.5              # Core Flask framework
Flask-RESTful==0.3.9      # For building REST APIs
Flask-Cors==3.0.10        # Enable CORS for frontend-backend communication
Flask-SQLAlchemy==3.0.5   # ORM for PostgreSQL
Flask-Migrate==4.0.4      # For Migration
psycopg2-binary==2.9.7    # PostgreSQL adapter for Python
python-dotenv==1.0.0      # Load environment variables from .env
marshmallow==3.20.1       # For data validation and serialization
marshmallow-sqlalchemy==0.28.0 # Integration of Marshmallow with SQLAlchemy

# # For real-time features (if using WebSocket):
# flask-socketio==5.3.2     # WebSocket support for Flask
# eventlet==0.33.3          # Required for Flask-SocketIO

# Deployment and security
gunicorn==20.1.0          # Production WSGI server for deploying the app
# Code Formatter
black==24.8.0